refactor: #97-updateMember 수정 - 조건 StringUtils.hasText(oldFile) 추가

+++ src/main/java/freshtrash/freshtrashbackend/controller/MemberApi.java
@@ -6,6 +6,7 @@ import freshtrash.freshtrashbackend.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -42,7 +43,7 @@ public class MemberApi {
                 MemberResponse.fromEntity(memberService.updateMember(memberPrincipal.id(), memberRequest, imgFile));
 
         // 파일이 수정된 경우 -> 이전 파일 삭제
-        if (!oldFile.equals(memberResponse.fileName())) {
+        if (StringUtils.hasText(oldFile) && !oldFile.equals(memberResponse.fileName())) {
             memberService.deleteOldFile(oldFile);
         }
         return ResponseEntity.ok(memberResponse);
